Proper ID3v1 support (on the backend level at least; GUI needs some amendments, mainly Genre field selection).
APE and ID3v1 tags are permitted to coexist in the same
MP3 file (and the APE tagging logic has been amended to
work around an existing ID3v1 tag). This coexistence is
not permitted for other formats, however the backend is
there, should the 'politics' change.

An ID3v1 tag is always at the very end of the file,
after the APE tag, if there is one.

Enhanced error reporting for all meta_write() functions.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@796 3cd24cdc-1f22-0410-b8b1-dcf80e670293