Initial incomplete implementation of ID3v2.
ID3v2.3 and ID3v2.4 will be supported to the
extent that is useful for Aqualung's purposes.

Currently reading of text frames (T***) including
TXXX (user defined text frame) are supported.

The following is still missing for complete
(as needed) read support of tags:
* COMM (Comments)
* APIC (Attached Picture frame)
* GEOB (General embedded file object)

Write support is completely missing ATM.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@836 3cd24cdc-1f22-0410-b8b1-dcf80e670293