Implement smarter way of transferring tags when creating (ripping or exporting) files: * encoder support for the <<meta>> structure, throwing out older, obsolete method * when exporting, create a new <<meta>> structure based on the existing <<meta>>   in the source file, and the knowledge of what tags are supported by the output format * Settings/Metadata: tag update options now also control what tags are added to newly   created files; update dialog text accordingly.
The new feature makes it possible to transcode files with intelligent metadata transfer.
For example, exporting a FLAC containing Ogg Vorbis comments and two FLAC picture frames
to MP3 will result in an MP3 with an ID3v2 having the respective text frames, and two
APIC frame for the pictures. It is also possible to set what combination of ID3v1,
ID3v2 and APE tags should be used when creating MPEG audio files -- the Settings/Metadata
frame used for batch update now controls this as well.
 


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@893 3cd24cdc-1f22-0410-b8b1-dcf80e670293