Better rollback operation when pausing playback.
When Aqualung is paused, internal sound buffers are flushed and the position is rolled back so that the audio that was present in the buffers is not skipped, but played when playback is resumed. This is necessary because we don't want to wait for the buffers to empty the natural way when being paused (this would cause a ~1.5 s lag between pressing pause button and the actual stopping of the audio).

However, there is also audio in the soundcard buffer (audio that we have already sent to the driver, but has not been actually played). We also have to compensate for this.

This commit adds support for reporting back the soundcard latency from the output threads to the disk thread, and taking this value into account when computing the amount of rollback. Since the rollback issue (skipping a portion of audio on pause/resume) was only prevalent on Win32, only the win32 output driver reports nonzero latency -- the others are precise enough without this compensation.



git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@933 3cd24cdc-1f22-0410-b8b1-dcf80e670293