Add support for sndio backend
This commit adds support for a sndio backend to Aqualung, bringing
the total number of backends to 5.  libsndio was recently introduced
in OpenBSD as a simple audio API that supports OpenBSD's builtin
sound server, aucat.

As libsndio is currently OpenBSD specific, sndio support is disabled
by default in configure, and needs to be enabled using --with-sndio.
This may be changed in the future if other operating systems add
support for it.

You can specify the sndio backend using -o sndio.  While libsndio
supports using a specified device, it is recommended that no device
be specified.  The sndio support in Aqualung does not allow
specifying a device, though you can control which device is used via
the AUDIODEVICE environment variable (which libsndio uses if it is
present).


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1034 3cd24cdc-1f22-0410-b8b1-dcf80e670293