Add programmable title format support
This commit embeds a Lua interpreter inside of Aqualung for the
purpose of allowing the user full control over the title format.  It
allows the user to use any metadata field that Aqualung recognizes,
as well as a few fields from the file info in order to compose a title
field.  See the documentation update included in this commit for
usage information for this feature.

Because of the way Aqualung is structured, the only easy place to add this
was inside the playlist_filemeta_get, since that's the only place you
have a playlist_data_t and the corresponding file_decoder_t.

Because this support is optional, and the user could have any number
of errors in their Lua code, Aqualung falls back to the standard
title formatting code if there are any issues using the programmable
title formatting code.

Because we don't want to store all of the metadata in the playlist,
we only store the formatted title string.  Since this is just a fixed
string and the playlist file doesn't contain the information to
recreate it if the programmable format changes, the user needs to
reread all file metadata if they change the programmable title format.

This commit also makes some other small changes.  It is a little more
paranoid in regards to making sure that pointers inside
playlist_data_t structures are NULL unless they point to valid memory.
It also updates the output of -v to include podcast support.  It
refactors metadata_get_* to call metadata_get_string_field.  It also
fixes some spacing in --configure help.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1055 3cd24cdc-1f22-0410-b8b1-dcf80e670293