Minor fix to sndio support
This results from discussion with the libsndio developers.  Instead of
using a fixed 1024-frame buffer, use a 250ms buffer.  This decreases
audio stutterring for me when I load a large playlist over NFS.  The
use of round is a because having a buffer that is twice the size of
round gives optimal performance.  The change of bufsz to appbufsz is
due to the fact that bufsz is now used slightly differently in
current versions of libsndio (it's now for calculating latency), and
it is read only.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1060 3cd24cdc-1f22-0410-b8b1-dcf80e670293