Don't require restart to update programmable title format file
This makes it much easier to modify and debug a programmable
title format.  When the Settings page is saved, the programmable
title format file is loaded (or reloaded if it hasn't changed).
Changes take effect immediately, but just as before, existing
entries on the playlist are not modified until their metadata
is read, so you need to use the Reread file metadata command
to change existing entries.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1095 3cd24cdc-1f22-0410-b8b1-dcf80e670293