Add support for an application_title lua function separate from the playlist_title lua function
The application_title lua function is used for the window's title
and for the main title label of the player.  The purpose of having
it distinct from the playlist_title function is for complicated
playlist titles that use a quasi-columnar format, where the format
for the playlist is not the desired window title.

On the lua end, the changes are fairly simple, simply add an
application_title global function to your .lua programmable title
format file, and Aqualung will now use it for the application titles.

This change is fully backwards compatible, as Aqualung will try
playlist_title if there is an error with application_title (but it
will log a message to stderr).

The GUI implementation requires some changes that appear necessary
with the current codebase.  For reasons I'm not sure about, when
Aqualung loads a new file, it refresh_displays once, then
process_metablock once, then refresh_displays twice (both
refresh_displays and process_metablock call set_title_label).  My
guess is this is due to unrelated information (other than the
title) being updated by the subsequent calls in refresh_displays,
and this commit modifies that behavior.  This commit disables
refresh_display's updating of the title label if programmable
title formatting is used.  The reason I chose this route is that
refresh_displays doesn't have knowledge of the fdec (file decoder)
for the current file, but process_metablock does, and you need the
fdec in order to use the programmable title formatting functions.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1096 3cd24cdc-1f22-0410-b8b1-dcf80e670293