Applying stop_after_current.diff by cobines:
Since rev. 1096 the [STOPPING] message is no longer shown in the title
bar if one is using LUA title formatting. The application title is now
remembered and is used when updating the title in toggle_stop_after_current_song().

In the patch was also added clearing the stop_after_current flag when
choosing previous/next song.


git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1117 3cd24cdc-1f22-0410-b8b1-dcf80e670293