Add support for custom playlist menu commands, using add_playlist_menu_command
This is the first step in adding a more general extension layer to
Aqualung, beyond simply formatting titles.  This commit adds the
ability to define custom playlist menu commands (and submenus),
with callbacks that call user-written Lua functions.

The basics are described in the documentation patch included in this
commit.

To make this and future extension patches easier, I've added
default functions to ext_title_format.c in the cases where HAVE_LUA
is not defined.  This way, such functions can be called in other
parts of the code without having to surround them with an
ifdef HAVE_LUA.

The attached patch is reflective of the general strategy for adding
extension code, which is to write as much of it as possible in Lua
itself.  It probably should be moved out of a .c file and into it's
own file that is statically compiled into the source, but that can
be done later.

Currently, the code lacks error checking when retrieving information
from Lua, and it's probably possible to crash Aqualung given a poorly
coded .lua file.  This will be fixed.

In order for extensions to be generally useful, the entire Lua
standard library is now added to the embedded interpreter.
Previously, only a few libraries were loaded.

Currently, the custom playlist menu gui is only created at startup
so you need to restart Aqualung for changes to take effect.
Hopefully this can be fixed without much difficulty.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1119 3cd24cdc-1f22-0410-b8b1-dcf80e670293