Add ability to define custom remote commands in Lua
See the included documentation for usage and an example.  Currently,
this isn't useful for much, since we haven't added any functions to
Lua that the user can call to affect Aqualung's behavior, or even
get information from Aqualung.

Remote functions defined by the user currently must accept 0
arguments.  This limits the utility, but in order to work around it
we'd need to decide on a protocol to use to pass the arguments over
the socket to Aqualung.

This commit also partly fixes the reloading of custom playlist menu
commands.  Previously, saving the settings page (which reloads the
Lua interpreter) caused all custom menu settings not to work (and
potentially a SEGV).  This avoids that but causes a memory leak
instead.  It is still not possible to add new menu entries
dynamically (or remove those that are no longer defined), and
fixing that will require some work and potentially a UI change.

This also adds a default case when receiving remote commands, which
prints an error message if an unsupported command type is received.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1121 3cd24cdc-1f22-0410-b8b1-dcf80e670293