Teardown and rebuild custom command menu when reloading the settings
Previously, the custom playlist command menu entries were fixed on
startup, the entries were added directly to the playlist menu.  This
commit makes it so that custom playlist commands are always on a
submenu, and that submenu is torn down and rebuilt every time the
lua interpreter is reloaded (when the settings window is saved).

This commit refactors Aqualung.process_playlist_menu into itself
and Aqualung.build_playlist_menu.  process_playlist_menu will
return true if there are custom playlist commands and false
otherwise, and no longer takes an argument.  build_playlist_menu
takes a lightuserdata argument and calls build_menu with it.

The basic strategy for rebuilding the menu is to keep the main
playlist menu entry for the custom playlist command submenu hidden
by default.  If there are custom commands, they are added to the
submenu and the main playlist menu entry is shown.  Everytime the
settings are reloaded, the submenu is destroyed, and rebuilt based
on the reloaded .lua file.

Reloading the settings still leaks a small amount of memory, but
hopefully that will be fixed soon.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1123 3cd24cdc-1f22-0410-b8b1-dcf80e670293