Don't leak memory when reloading the Lua interpreter
Before, reloading the Lua interpreter leaked memory for each
custom playlist menu entry, because it has to strdup the strings
it received from Lua before using them as the pointer arguments in
GTK callbacks, but it wasn't keeping track of them.

This commit keeps track of them in the Lua registry, and frees them
before closing the interpreter.

It also adds a prototype for add_custom_commands_to_playlist_menu
to fix a compiler warning.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1124 3cd24cdc-1f22-0410-b8b1-dcf80e670293