Add support for custom keybindings in the main window
This commit adds complete support for custom keybindings to the
main window, as well as the ground work for supporting custom
keybindings in the playlist and music store windows.  I don't use
the music store and I embed the playlist in the main window, so
I haven't hooked into the keybindings of either of those two
separately.  It should be fairly simple add support similar to
how the main window keybindings are handled.

The custom keybinding support is started with the Ctrl-z or
Ctrl-Z keys.  When that key combination is received, it signals
that the next keypress is a custom keybinding.  The custom
keybinding can be any keyboard key, as well as any of the shift,
control, alt (Mod1), and super (Mod4/Windows) keys.  See the
included documentation patch for more details.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1126 3cd24cdc-1f22-0410-b8b1-dcf80e670293