No longer provide an argument to the custom menu callback function
Previously, the callback received an array of selected files for
the playlist.  However, it's possible that the menu callback
doesn't need such a list.  Inside, have the menu callback accept
0 arguments, and if an array of selected filenames is desired,
the extension writer can call selected_files(), a global method
that gives the same results as the argument previously given to
the menu callback.  This has the advantage of making the array
of selected filenames available both to custom remote commands
as well as custom keybindings.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1127 3cd24cdc-1f22-0410-b8b1-dcf80e670293