Add support for hooks for executing extension code when certain events occur.
This is the final piece of the extension support foundation.  After
this, it's just adding additional functions/hooks/etc, not new types
of extensions.

Currently, the only hook supported is track_change, which fires when
a new track is cued up.  See the included documentation patch for
an example.

While here, fix the error message for errors when running a
custom keybinding function.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1129 3cd24cdc-1f22-0410-b8b1-dcf80e670293