Minor cleanup of files to use Lua extension support/file as the official setting
No restructuring of the options GUI, yet.  This changes the configure
output, about box, and aqualung -v output.

While here, rename the ext_title_format.* files to ext_lua.*.

Also, update the documentation to move the programmable title format
information to the end with the rest of the extension information.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1139 3cd24cdc-1f22-0410-b8b1-dcf80e670293