Some GUI tweaking, don't reload Lua interpreter on every settings OK
In line with Tom's suggestions, this moves the Title format stuff
onto the main General Settings page, and changes the settings to
not reload the Lua interpreter every time the settings page is
closed.  It now only reloads the interpreter if the file name
has changed.

To reload the Lua interpreter manually, a playlist menu entry
is added.  Some other minor restructuring of the playlist menu
is also done.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1140 3cd24cdc-1f22-0410-b8b1-dcf80e670293