automake: removed ancient automake helpers
Aqualung actually builds out nicely with autoreconf, so just use that
and let it install the maintainer's [modern] versions of the helpers.
(Distribution archives should still include all the helpers and a
generated configure script, but we don't need to keep them in version
control.) Stop explicitly checking for aclocal and autoheader, it's
enough to check for autoconf and let autoreconf invoke them as
necessary.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1174 3cd24cdc-1f22-0410-b8b1-dcf80e670293