autopoint: refactor GNU gettext integration
Instead of using home-brew automake to invoke gettext, use autopoint
to install all the helper macros and templates for us.  By using the
gettext/autopoint autoconf hooks we get built-in checking for the
necessary tools complete with version restrictions when autoreconf is
run.  Move all the po files to the top of the build tree to match the
traditional GNU project layout.  Note that source files with strings
requiring translation need to be explicitly listed in po/POTFILES.in
now.  Additionally, by using the gettext helper macros, we can now
take advantage of the LIBINTL variable for inclusion of any necessary
linker flags for the host platform, meaning we don't have to hard-code
assumptions like we used to for win32 builds.

The inclusion of the gettext.h convenience header (which allows folks
to build Aqualung without NLS if they want to) has been done here
without involving gnulib because while the gettext-h gnulib module is
LGPLv2+ and hence compatible with Aqualung's own GPLv2+ license,
modern gnulib infrastructure tends to be GPLv3+ and I didn't want to
cross that bridge just yet.  Hence, the somewhat sloppy approach of
just copying gettext.h to the top of the build tree (and knowing that
it will be found because autoconf's config.h lives there).

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1175 3cd24cdc-1f22-0410-b8b1-dcf80e670293