autoconf: introduce custom macro AQUALUNG_DETECT
AQUALUNG_DETECT(package, brief-description, [detect-actions])

Wrapper around AC_ARG_WITH for setting up autodetection of package.
Establishes a --with-package / --without-package configure script
option, with the $with_package variable set to "detect" by default,
"yes" if --with-package is given, or "no" if --without-package is
given.  If $with_package is "detect" or "yes" then any given
detect-actions are run, with the expectation they will set
$with_package accordingly if the optional package is found usable or
not.  (The AQUALUNG_FOUND and AQUALUNG_MISSING macros are fashioned
for that purpose.)  After any detect-actions are run, if $with_package
is still "detect" then then $with_package is set to "not found", but
if $with_package starts with "yes" then an AC_DEFINE definition of
HAVE_package (with package transliterated to all uppercase) is
defined.  To support conditional compilation scenarios, an automake
conditional with the same name as the AC_DEFINE definition is
established, it will evaluate to true if $with_package starts with
"yes" and false otherwise.  The brief-description given is used for
the configure script help text, an AC_MSG_NOTICE at run-time, and the
comments in the AC_CONFIG_HEADERS definitions.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1215 3cd24cdc-1f22-0410-b8b1-dcf80e670293