libdecoder: use conditional compilation
Use automake's conditionals to only compile the glue for a codec when
its decoder library is available.  This makes the resulting binary a
tiny bit smaller, means we don't need C++ at all unless Aqualung is
configured with Monkey's Audio Codec support, and means the codec glue
doesn't need to include stub routines for the "not used" case anymore.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1217 3cd24cdc-1f22-0410-b8b1-dcf80e670293