includes: clean up libcdio and Monkey's Audio Codec headers
Both libcdio and Monkey's Audio Codec enjoy the dubious practice of
leaking their autoconf generated macro values into the global
namespace.  This can cause all sorts of irritating problems, but the
least subtle collisions are the PACKAGE* and VERSION macros which are
always going to be different from those of our app and hence cause
the preprocessor to spew redefinition warnings.  So, to work around
the braindamage present in these libraries, we take the approach of:

  1) undefine any macros that could be present in our own config.h
     that could legitimately be defined differently in the buggy lib
  2) include the headers of the buggy lib
  3) undefine all the same macros as #1 again
  4) re-include our config.h; thus headers included after the buggy
     lib and our code will have macros expanded correctly

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1244 3cd24cdc-1f22-0410-b8b1-dcf80e670293