Allow track_change hook to access m and i playlist formatting functions
This adds a run_hooks_fdec function for running hooks with a
passed file_decoder_t*, allowing the lua function inside the hook
to access the metadata for the file.  It moves the run_hooks call
to the place with access to the file_decoder_t* when the track is
changed.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1263 3cd24cdc-1f22-0410-b8b1-dcf80e670293