Ease debugging in lua extension code
This makes sure that if there is a bug reported in the prelude code,
lua will include line number information for it and make it
obvious that it occurs in the prelude (AQUALUNG_LUA_API).

When attempting to add a hook for an invalid hook type, use a
descriptive error message, and make sure the error information
points to the line in the user's extension file and not the
prelude code.

git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@1266 3cd24cdc-1f22-0410-b8b1-dcf80e670293