Refactor file info dialog invocation and access to tree stores
This is needed to create a feature allowing to go to the previous /
next file in the natural order of wherever the file info dialog was
opened from, without closing and reopening the dialog.

The prev/next buttons to do this are already on the file info dialog,
but they don't yet rebuild the dialog content. However, the generic
tree traversal is already implemented, taking care that the containing
store (in case the dialog was opened from a file or podcast store)
cannot be left while browsing.

git-svn-id: https://svn.code.sf.net/p/aqualung/code/trunk@1293 3cd24cdc-1f22-0410-b8b1-dcf80e670293