Avoid double free of send_button and add_tag_table.
These two widgets are created dynamically, but they're added to the info
window's widget structure so they will be destroyed automatically by GTK
when the window is destroyed.

Since the window close callback also calls fi_unload, this resulted in
the widgets being destroyed twice. NULL out the pointers in the callback
to prevent this.