Fix crash when closing both Search dialogs
With Gtk3 there is some weird interaction that causes a crash when
either of the Search dialogs (for searching in the Music Store and the
Playlist) is closed with a selection in the list. This has definitely
not crashed before moving to Gtk3 and debugging it reveals some weird
issue that looks like the selection_changed handler re-triggering on
gtk_list_store_clear() even after issuing an unselect_all() on the
selection before closing the dialog. So we revert to a brute force
solution to paper over this bug in Gtk3 or elsewhere...